requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.19.0
PyPDF2>=3.0.0
Flask>=2.0.0
Werkzeug>=2.0.0 